Deepen metric pipeline: merge 3 shallow modules into snapshot.py
Collapse _raw_metrics.py, _dev_repo_metrics.py, and _row_factory.py into
snapshot.py. These modules were pass-through wiring layers — their
interfaces were as complex as their implementations, serving a single
consumer.

- RawMetrics dataclass moved to _rows.py (leaf module, avoids circular dep)
- compute_raw, compute_dev_metrics, compute_repo_metrics, raw_to_row,
  rank_rows moved to snapshot.py (private)
- band_from_health kept public, exported via __init__.py
- health.py imports RawMetrics from _rows.py now
- Deleted test_dev_repo_metrics.py and test_row_factory.py (coverage
  subsumed by existing test_snapshot.py)

Retained as separate modules: calculator.py (reused by trend pipeline),
health.py (114 lines of domain logic), _ai_detection.py (reused), _rows.py
(leaf type defs).

ADR-0006 records the boundary decisions so future architecture reviews
don't re-suggest splitting these differently.
