+++ git_dev_metrics/metrics/__init__.py
@@ -1,6 +1,5 @@
 """Metrics"""
 
-from ._row_factory import band_from_health
 from ._rows import Band, Row, Summary, band_color
 from .calculator import (
     calculate_avg_lines_per_pr,
@@ -15,7 +14,7 @@ from .calculator import (
 from .health import calculate_dev_health_score, calculate_health_score
 from .loader import InvalidRangeError, load_snapshot_for_months, load_snapshot_for_range
 from .printer import ConsolePrinter, Printer
-from .snapshot import MetricsSnapshot, build_summary
+from .snapshot import MetricsSnapshot, band_from_health, build_summary
 
 __all__ = [
     "Band",
